requirements.txt
asgiref==3.7.2
certifi==2023.7.22
distlib==0.3.7
Django==4.2.5
django-crispy-forms==2.0
filelock==3.12.3
pipenv==2023.9.1
platformdirs==3.10.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.4
django-cors-headers==4.3.0
djangorestframework==3.14.0
crispy-bootstrap4==2023.1
requests==2.31.0
openai==0.28.1